+++ cocos/physics/bullet/shapes/bullet-simplex-shape.ts
@@ -33,7 +33,7 @@ import { SimplexCollider } from '../../../../exports/physics-framework';
 import { cocos2BulletVec3 } from '../bullet-utils';
 import { ISimplexShape } from '../../spec/i-physics-shape';
 import { IVec3Like } from '../../../core/math/type-define';
-import { bt } from '../bullet.asmjs';
+import { bt } from '../instantiated';
 import { BulletCache } from '../bullet-cache';
 
 export class BulletSimplexShape extends BulletShape implements ISimplexShape {
