+++ cocos/physics/bullet/shapes/bullet-sphere-shape.ts
@@ -33,7 +33,7 @@ import { PhysicsSystem, SphereCollider } from '../../../../exports/physics-frame
 import { cocos2BulletVec3 } from '../bullet-utils';
 import { ISphereShape } from '../../spec/i-physics-shape';
 import { BulletCache, CC_V3_0 } from '../bullet-cache';
-import { bt } from '../bullet.asmjs';
+import { bt } from '../instantiated';
 import { absMaxComponent } from '../../../core';
 
 export class BulletSphereShape extends BulletShape implements ISphereShape {
